Laravel 7 support
