+++ .github/workflows/test.yml
@@ -15,7 +15,7 @@ jobs:
       fail-fast: true
       matrix:
         php: [7.2, 7.3, 7.4]
-        laravel: [5.6.*, 5.7.*, 5.8.*, 6.*]
+        laravel: [5.6.*, 5.7.*, 5.8.*, 6.*, 7.*]
         dependency-version: [prefer-lowest, prefer-stable]
         include:
           - laravel: 5.6.*
@@ -26,6 +26,8 @@ jobs:
             testbench: 3.8.*
           - laravel: 6.*
             testbench: 4.*
+          - laravel: 7.*
+            testbench: 5.*
         exclude:
           - php: 7.4
             laravel: 5.6.*
