+++ composer.json
@@ -9,20 +9,16 @@
         }
     ],
     "require": {
-        "illuminate/database": "~5.6.0|~5.7.0|~5.8.0|^6.0",
-        "illuminate/support": "~5.6.0|~5.7.0|~5.8.0|^6.0",
-        "illuminate/validation": "~5.6.0|~5.7.0|~5.8.0|^6.0",
-        "symfony/http-kernel": "^4.0"
+        "illuminate/database": "~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0",
+        "illuminate/support": "~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0",
+        "illuminate/validation": "~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0",
+        "symfony/http-kernel": "^4.0|^5.0"
     },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Livewire\\LivewireServiceProvider"
-            ],
-            "aliases": {
-                "Livewire": "Livewire\\Livewire"
-            }
-        }
+    "require-dev": {
+        "phpunit/phpunit": "^7.5.15|^8.4|^9.0",
+        "laravel/framework": "~5.6.0|~5.7.0|~5.8.0|^6.0|^7.0",
+        "orchestra/testbench": "~3.6.0|~3.7.0|~3.8.0|^4.0|^5.0",
+        "psy/psysh": "@stable"
     },
     "autoload": {
         "psr-4": {
@@ -35,10 +31,19 @@
             "App\\": "vendor/orchestra/testbench-core/laravel/app"
         }
     },
-    "require-dev": {
-        "phpunit/phpunit": "^7.5.15|^8.4",
-        "laravel/framework": "~5.6.0|~5.7.0|~5.8.0|^6.0",
-        "orchestra/testbench": "^3.8|^4.0",
-        "psy/psysh": "@stable"
-    }
+    "extra": {
+        "laravel": {
+            "providers": [
+                "Livewire\\LivewireServiceProvider"
+            ],
+            "aliases": {
+                "Livewire": "Livewire\\Livewire"
+            }
+        }
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
